perf(SubscriptionUserinfo): use `i128`

+++ src/model/mod.rs
@@ -5,8 +5,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize)]
 pub struct SubscriptionUserinfo {
-    pub upload: isize,
-    pub download: isize,
-    pub total: isize,
-    pub expire: isize,
+    pub upload: i128,
+    pub download: i128,
+    pub total: i128,
+    pub expire: i128,
 }
